[Refactor] Custom TextField 재구성
